% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{annotation_data}
\alias{annotation_data}
\title{GENCODE v36 Gene Annotations}
\format{
A data frame with 67,016 rows and 2 columns:
\describe{
\item{gene_id}{ENSEMBL gene ID (e.g., "ENSG00000223972")}
\item{gene_name}{Official gene symbol (e.g., "DDX11L1")}
}
}
\source{
Subset from gencode.v36.annotation.gtf.gz
}
\usage{
annotation_data
}
\description{
Gene ID to gene name mappings from GENCODE release 36 (GRCh38.p13).
}
\references{
\url{https://www.gencodegenes.org/human/release_36.html}
}
\keyword{datasets}
