% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkData.R
\name{checkData}
\alias{checkData}
\title{Check Data}
\usage{
checkData(lacenObject, orgdb = "org.Hs.eg.db")
}
\arguments{
\item{lacenObject}{An object of class "lacen" created by \code{\link[=initLacen]{initLacen()}}.}

\item{orgdb}{A Bioconductor \code{OrgDb} object corresponding to the organism of interest (e.g., \code{org.Hs.eg.db} for human).}
}
\value{
Prints a confirmation message if data format is correct; otherwise, issues warnings.
}
\description{
Validates the format of data within a lacen object.
}
