% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterTransform.R
\name{filterTransform}
\alias{filterTransform}
\title{Filter and Transform}
\usage{
filterTransform(
  lacenObject,
  pThreshold = 0.01,
  fcThreshold = 1,
  topVarGenes = 5000,
  filterMethod = "DEG",
  ...
)
}
\arguments{
\item{lacenObject}{An object of class "lacen" created by \code{\link[=initLacen]{initLacen()}}.}

\item{pThreshold}{p-value threshold for differential expression.}

\item{fcThreshold}{Fold change threshold for differential expression.}

\item{topVarGenes}{Number of top variable genes to retain if filtering by variance.}

\item{filterMethod}{Method to filter genes: "DEG" (default) or "var".}
}
\value{
A 'lacen' S3 object with updated \code{datExpr}.
}
\description{
Removes genes with low variation, filters expression data by MAD or DEG, and applies voom correction.
}
