% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixData.R
\name{fixData}
\alias{fixData}
\title{Process and Format Input Data for a lacen Object}
\usage{
fixData(
  annotationData = annotation_data,
  datCounts = rawExpressionData,
  datExpression = expressionDGEData,
  datTraits = metadata,
  ncAnnotation = ncAnnotation,
  verbose = TRUE
)
}
\arguments{
\item{annotationData}{Annotation dataframe with "gene_id" and "gene_name" columns.}

\item{datCounts}{Data counts dataframe with genes as rows and samples as columns.
The gene/transcript IDs should be in the row names and sample names/IDs in the column names.}

\item{datExpression}{Differential expression dataframe. Can be from Limma output or a dataframe with columns "ID", "log2FC", and "pvalue".}

\item{datTraits}{Conditions/Traits data. A two-column dataframe with sample IDs in "Sample" and condition codes in "Trait".}

\item{ncAnnotation}{Subset of "annotationData" containing only long non-coding RNAs.}

\item{verbose}{A logical value (\code{TRUE} or \code{FALSE}) indicating whether to print
progress messages to the console. Defaults to \code{TRUE}.}
}
\value{
A 'lacen' S3 object containing the cleaned,
formatted, and validated data, ready for further analysis.
}
\description{
This function takes raw count data, differential expression results, sample
traits, and annotation data, then processes and formats them into a
standardized \code{lacen} object for downstream analysis.
}
\details{
During processing, genes may be filtered out if they are not present in the
annotation data. Additionally, the column names of the input data frames
may be standardized to ensure consistency.
}
