% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapTopConnectivity.R
\name{heatmapTopConnectivity}
\alias{heatmapTopConnectivity}
\title{Heatmap of Top Connectivity}
\usage{
heatmapTopConnectivity(
  lacenObject,
  module,
  submodule = FALSE,
  hmDimensions = FALSE,
  filename = FALSE,
  removeNonDEG = FALSE,
  outTSV = FALSE,
  ...
)
}
\arguments{
\item{lacenObject}{An object of class "lacen" created by \code{\link[=initLacen]{initLacen()}}.}

\item{module}{Numeric value indicating the module to analyze.}

\item{submodule}{Optional. Numeric value indicating the submodule to analyze. Set to FALSE to ignore.}

\item{hmDimensions}{Numeric value indicating the number of genes to display in the heatmap.}

\item{filename}{Filename to save the heatmap as PNG.}

\item{removeNonDEG}{Logical. If TRUE, displays only differentially expressed genes.}

\item{outTSV}{Logical or character string. If FALSE, does not save. If a path is provided, saves the TSV with heatmap values.}

\item{...}{Additional parameters for internal functions.}
}
\value{
None (plots and saves the heatmap).
}
\description{
Generates a heatmap of a specified module or submodule, showing interconnectedness based on the TOM.
}
