% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initLacen.R
\name{initLacen}
\alias{initLacen}
\title{Lacen Constructor}
\usage{
initLacen(
  datCounts = NULL,
  datExpr = NULL,
  datExpression = NULL,
  datTraits = NULL,
  annotationData = NULL,
  ncAnnotation = NULL,
  keepSamples = NULL,
  height = NULL,
  indicePower = NULL,
  modGroups = NULL,
  bootstrapStability = FALSE,
  bootstrap = FALSE,
  cutBootstrap = FALSE,
  summdf = NULL,
  rrvgolist = NULL,
  TOM = NULL
)
}
\arguments{
\item{datCounts}{Data counts dataframe with genes as rows and samples as columns.
The gene/transcript IDs should be in the row names and sample names/IDs in the column names.}

\item{datExpr}{Filtered and transformed expression data obtained via \code{\link[=filterTransform]{filterTransform()}}.}

\item{datExpression}{Differential expression dataframe. Can be from Limma output or a dataframe with columns "ID", "log2FC", and "pvalue".}

\item{datTraits}{Conditions/Traits data. A two-column dataframe with sample IDs in "Sample" and condition codes in "Trait".}

\item{annotationData}{Annotation dataframe with "gene_id" and "gene_name" columns.}

\item{ncAnnotation}{Subset of "annotationData" containing only long non-coding RNAs.}

\item{keepSamples}{Boolean vector indicating samples to retain based on outlier removal.}

\item{height}{Numeric value to determine outlier samples from the cluster tree plot.}

\item{indicePower}{Soft threshold power used for network construction.}

\item{modGroups}{Module groups from bootstrap parameter set.}

\item{bootstrapStability}{Stability ratios from bootstrap iterations via \code{\link[=lacenBootstrap]{lacenBootstrap()}}.}

\item{bootstrap}{Bootstrap parameters set by \code{\link[=lacenBootstrap]{lacenBootstrap()}}.}

\item{cutBootstrap}{Threshold to discard low stability genes.}

\item{summdf}{Summarized results from \code{\link[=summarizeAndEnrichModules]{summarizeAndEnrichModules()}}.}

\item{rrvgolist}{Enriched and reduced terms from \code{\link[=summarizeAndEnrichModules]{summarizeAndEnrichModules()}}.}

\item{TOM}{Topological overlap matrix from \code{\link[=summarizeAndEnrichModules]{summarizeAndEnrichModules()}}.}
}
\value{
A 'lacen' S3 object containing all the provided parameters.
}
\description{
Create a new 'lacen' S3 object.
}
