% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lacenBootstrap.R
\name{lacenBootstrap}
\alias{lacenBootstrap}
\title{Bootstrap Stability}
\usage{
lacenBootstrap(
  lacenObject,
  numberOfIterations = 100,
  maxBlockSize = 50000,
  pathModGroupsPlot = "3_modgroupsnew.png",
  pathStabilityPlot = "4_stability_bootstrap.png",
  nThreads = 4,
  ...
)
}
\arguments{
\item{lacenObject}{An object of class "lacen" created by \code{\link[=initLacen]{initLacen()}}.}

\item{numberOfIterations}{Number of bootstrap iterations to perform.}

\item{maxBlockSize}{Maximum number of genes processed simultaneously.}

\item{pathModGroupsPlot}{Filename to save the module stability plot as PNG.}

\item{pathStabilityPlot}{Filename to save the stability ratio plot as PNG.}

\item{nThreads}{Number of threads to use for parallel processing.}

\item{...}{Additional parameters for internal functions.}
}
\value{
A 'lacen' S3 object with updated \code{bootstrapStability} and \code{bootstrap} data.
}
\description{
Repeats network construction multiple times with subsets of genes to assess gene stability.
}
