% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lncRNAEnrich.R
\name{lncRNAEnrich}
\alias{lncRNAEnrich}
\title{lncRNA Enrichment Analysis}
\usage{
lncRNAEnrich(
  lncName,
  lacenObject,
  nGenes = 100,
  nHighlight = 10,
  sources = c("GO:BP", "GO:MF", "GO:CC", "KEGG", "REAC"),
  organism = "hsapiens",
  nGenesNet = 10,
  nTerm = 3,
  ...
)
}
\arguments{
\item{lncName}{Gene name of the lncRNA for enrichment analysis.}

\item{lacenObject}{An object of class "lacen" created by \code{\link[=initLacen]{initLacen()}}.}

\item{nGenes}{Number of top correlated genes to include in the analysis.}

\item{nHighlight}{Number of top enriched terms to highlight in the enrichment graph.}

\item{sources}{Functional term sources for enrichment analysis.}

\item{organism}{Organism name for functional annotations.}

\item{nGenesNet}{Number of genes to visualize in the network plot.}

\item{nTerm}{Number of top enriched terms to display in the enrichment graph.}

\item{...}{Additional parameters for internal functions.}
}
\value{
None (plots and saves enrichment results and network).
}
\description{
Performs functional enrichment analysis for a specified lncRNA based on top correlated genes.
}
