% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadGTF.R
\name{loadGTF}
\alias{loadGTF}
\title{Load GTF File}
\usage{
loadGTF(file)
}
\arguments{
\item{file}{Path to the local GTF file.}
}
\value{
A dataframe with "gene_id" and "gene_name" columns.
}
\description{
Loads a local GTF annotation file and returns it as a dataframe.
}
