% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nc_annotation}
\alias{nc_annotation}
\title{Long Non-coding RNA Annotations}
\format{
A data frame with 17,946 rows and 2 columns:
\describe{
\item{gene_id}{ENSEMBL gene ID (e.g., "ENSG00000243485")}
\item{gene_name}{lncRNA symbol (e.g., "MIR1302-2HG")}
}
}
\source{
Filtered from gencode.v36.long_noncoding_RNAs.gtf.gz
}
\usage{
nc_annotation
}
\description{
Subset of GENCODE v36 annotations containing only lncRNA genes.
}
\keyword{datasets}
