% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSoftThreshold.R
\name{plotSoftThreshold}
\alias{plotSoftThreshold}
\title{Plot Soft Threshold}
\usage{
plotSoftThreshold(
  lacenObject,
  filename = "indicePower.png",
  maxBlockSize = 10000,
  plot = TRUE
)
}
\arguments{
\item{lacenObject}{An object of class "lacen" created by \code{\link[=initLacen]{initLacen()}}.}

\item{filename}{Filename to save the soft threshold plot as PNG.}

\item{maxBlockSize}{Maximum number of genes processed at the same time.}

\item{plot}{Logical. If TRUE, displays the plot in the R session.}
}
\value{
None (plots and saves the soft threshold plot).
}
\description{
Generates a plot to help select the soft-thresholding power for network construction.
}
