% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectSoftThreshold.R
\name{selectSoftThreshold}
\alias{selectSoftThreshold}
\title{Select Soft Threshold}
\usage{
selectSoftThreshold(lacenObject, indicePower)
}
\arguments{
\item{lacenObject}{An object of class "lacen" created by \code{\link[=initLacen]{initLacen()}}.}

\item{indicePower}{Numeric value indicating the selected soft threshold power.}
}
\value{
A 'lacen' S3 object with updated \code{indicePower}.
}
\description{
Sets the chosen soft threshold power in the lacen object.
}
