% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setBootstrap.R
\name{setBootstrap}
\alias{setBootstrap}
\title{Set Bootstrap Cut-off}
\usage{
setBootstrap(lacenObject, cutBootstrap)
}
\arguments{
\item{lacenObject}{An object of class "lacen" created by \code{\link[=initLacen]{initLacen()}}.}

\item{cutBootstrap}{Numeric value between 0 and 1 indicating the stability cutoff.}
}
\value{
A 'lacen' S3 object with updated \code{cutBootstrap}.
}
\description{
Sets a cutoff value for bootstrap stability to filter out less stable genes.
}
