% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackedBarplot.R
\name{stackedBarplot}
\alias{stackedBarplot}
\title{Module-Correlation Stacked Barplot}
\usage{
stackedBarplot(lacenObject, filename = "stackedplot_desk.png", plot = TRUE)
}
\arguments{
\item{lacenObject}{An object of class "lacen" created by \code{\link[=initLacen]{initLacen()}}.}

\item{filename}{Filename to save the stacked barplot as PNG.}

\item{plot}{Logical. If TRUE, displays the plot in the R session.}
}
\value{
None (plots and saves the stacked barplot).
}
\description{
Creates a barplot illustrating eigengene modules, their gene types, and trait-module correlations.
}
