% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeAndEnrichModules.R
\name{summarizeAndEnrichModules}
\alias{summarizeAndEnrichModules}
\title{Summarize and Enrich Modules}
\usage{
summarizeAndEnrichModules(
  lacenObject,
  maxBlockSize = 20000,
  TOMType = "unsigned",
  minModuleSize = 30,
  reassignThreshold = 0,
  mergeCutHeight = 0.3,
  pamRespectsDendro = FALSE,
  corType = "bicor",
  userThreshold = 0.05,
  sources = "BP",
  organism = "hsapiens",
  orgdb = "org.Hs.eg.db",
  reducedTermsThreshold = 0.7,
  filename = "5_enrichedgraph.png",
  pathTSV = FALSE,
  ...
)
}
\arguments{
\item{lacenObject}{An object of class "lacen" created by \code{\link[=initLacen]{initLacen()}}.}

\item{maxBlockSize}{Maximum number of genes processed simultaneously.}

\item{TOMType}{Type of Topological Overlap Matrix.}

\item{minModuleSize}{Minimum number of genes per module.}

\item{reassignThreshold}{p-value threshold for reassigning genes between modules.}

\item{mergeCutHeight}{Dendrogram cut height for merging modules.}

\item{pamRespectsDendro}{Logical. If TRUE, ensures PAM respects dendrogram structure.}

\item{corType}{Type of correlation to use: "pearson" or "bicor".}

\item{userThreshold}{p-value threshold for enrichment significance.}

\item{sources}{Functional term sources for enrichment analysis (e.g., "GO:BP").}

\item{organism}{Organism name (e.g., "hsapiens").}

\item{orgdb}{Bioconductor OrgDb package (e.g., "org.Hs.eg.db").}

\item{reducedTermsThreshold}{Similarity threshold for reducing enrichment terms.}

\item{filename}{Filename to save the enrichment graph as PNG.}

\item{pathTSV}{Path to save the tsv raw result of the analysis.}

\item{...}{Additional parameters for internal functions.}
}
\value{
A 'lacen' S3 object with updated \code{summdf}, \code{rrvgolist}, and \code{TOM}.
}
\description{
Constructs the co-expression network, performs module enrichment analysis, and summarizes the data.
}
