% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkData.R
\name{checkData}
\alias{checkData}
\title{Check Data}
\usage{
checkData(lacenObject)
}
\arguments{
\item{lacenObject}{An object of class "lacen" created by \code{\link[=initLacen]{initLacen()}}.}
}
\value{
Prints a confirmation message if data format is correct; otherwise, issues warnings.
}
\description{
Validates the format of data within a lacen object.
}
